Moving terminal back to public.
