+++ container/container_test.go
@@ -30,14 +30,16 @@ import (
 	"github.com/mum4k/termdash/internal/event/testevent"
 	"github.com/mum4k/termdash/internal/keyboard"
 	"github.com/mum4k/termdash/internal/mouse"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
 	"github.com/mum4k/termdash/internal/terminalapi"
 	"github.com/mum4k/termdash/internal/widgetapi"
+	"github.com/mum4k/termdash/terminal/faketerm"
+	"github.com/mum4k/termdash/widgets/barchart"
 	"github.com/mum4k/termdash/widgets/fakewidget"
 )
 
 // Example demonstrates how to use the Container API.
 func Example() {
+	bc, err := barchart.New()
 	if _, err := New(
 		/* terminal = */ nil,
 		SplitVertical(
@@ -61,7 +63,7 @@ func Example() {
 			),
 			Right(
 				Border(draw.LineStyleLight),
-				PlaceWidget(fakewidget.New(widgetapi.Options{})),
+				PlaceWidget(bc),
 			),
 		),
 	); err != nil {
