+++ container/traversal_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	"github.com/mum4k/termdash/internal/cell"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
+	"github.com/mum4k/termdash/terminal/faketerm"
 )
 
 func TestRoot(t *testing.T) {
