+++ internal/canvas/braille/braille_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/mum4k/termdash/internal/canvas"
 	"github.com/mum4k/termdash/internal/canvas/testcanvas"
 	"github.com/mum4k/termdash/internal/cell"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
+	"github.com/mum4k/termdash/terminal/faketerm"
 )
 
 func Example_copiedToCanvas() {
