+++ internal/canvas/testcanvas/testcanvas.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/mum4k/termdash/internal/canvas"
 	"github.com/mum4k/termdash/internal/cell"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
+	"github.com/mum4k/termdash/terminal/faketerm"
 )
 
 // MustNew returns a new canvas or panics.
