+++ internal/draw/segdisp/segment/segment_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/mum4k/termdash/internal/cell"
 	"github.com/mum4k/termdash/internal/draw"
 	"github.com/mum4k/termdash/internal/draw/testdraw"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
+	"github.com/mum4k/termdash/terminal/faketerm"
 )
 
 func TestHV(t *testing.T) {
