+++ terminal/faketerm/diff.go
