+++ terminal/faketerm/faketerm.go
