+++ terminal/termbox/cell_options.go
