+++ terminal/termbox/cell_options_test.go
