+++ terminal/termbox/color_mode.go
