+++ widgets/fakewidget/fakewidget_test.go
@@ -24,9 +24,9 @@ import (
 	"github.com/mum4k/termdash/internal/draw/testdraw"
 	"github.com/mum4k/termdash/internal/keyboard"
 	"github.com/mum4k/termdash/internal/mouse"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
 	"github.com/mum4k/termdash/internal/terminalapi"
 	"github.com/mum4k/termdash/internal/widgetapi"
+	"github.com/mum4k/termdash/terminal/faketerm"
 )
 
 // keyEvents are keyboard events to send to the widget.
