+++ widgets/segmentdisplay/segmentdisplaydemo/segmentdisplaydemo.go
@@ -24,8 +24,8 @@ import (
 	"github.com/mum4k/termdash/container"
 	"github.com/mum4k/termdash/internal/cell"
 	"github.com/mum4k/termdash/internal/draw"
-	"github.com/mum4k/termdash/internal/terminal/termbox"
 	"github.com/mum4k/termdash/internal/terminalapi"
+	"github.com/mum4k/termdash/terminal/termbox"
 	"github.com/mum4k/termdash/widgets/segmentdisplay"
 )
