+++ widgets/text/line_trim_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/mum4k/termdash/internal/canvas"
 	"github.com/mum4k/termdash/internal/canvas/testcanvas"
 	"github.com/mum4k/termdash/internal/draw/testdraw"
-	"github.com/mum4k/termdash/internal/terminal/faketerm"
+	"github.com/mum4k/termdash/terminal/faketerm"
 )
 
 func TestLineTrim(t *testing.T) {
